+++ CHANGELOG.md
@@ -16,6 +16,9 @@
 * Remove reason and servername label from tsm_status
 * Make tsm_db_buffer_hit_ratio and tsm_db_pkg_hit_ratio a ratio between 0.0-1.0
 * Rename tsm_db_buffer_total_requests to tsm_db_buffer_requests_total
+* Rename tsm_db_last_backup_time to tsm_db_last_backup_timestamp_seconds
+* Rename tsm_replication_end_time to tsm_replication_end_timestamp_seconds
+* Rename tsm_replication_start_time to tsm_replication_start_timestamp_seconds
 
 ### Changes